设备在线离线 首页事件count

+++ youlai-boot-master/src/main/java/com/youlai/boot/dashBoard/model/vo/DashCount.java
@@ -19,5 +19,5 @@ public class DashCount {
     private Long roomCount;
     //todo
     private Long demo1Count;
-    private Integer demo2Count;
+    private Long demo2Count;
 }
